+++ Packages/RimeBridge/Sources/RimeBridge/RimeBridgeCapabilities.swift
@@ -1,0 +1,12 @@
+import Foundation
+import RimeBridgeObjC
+
+public enum RimeBridgeCapabilities {
+    public static var luaModuleCompiledIn: Bool {
+        RimeDeployer.luaModuleCompiledIn()
+    }
+
+    public static var deploymentModules: [String] {
+        RimeDeployer.configuredModules().map { String(describing: $0) }
+    }
+}